add sync as an option
