+++ src/main/java/org/corfudb/infrastructure/CorfuServer.java
@@ -60,12 +60,13 @@ public class CorfuServer {
             "Corfu Server, the server for the Corfu Infrastructure.\n"
             + "\n"
             + "Usage:\n"
-            + "\tcorfu_server (-l <path>|-m) [-s] [-a <address>] [-t <token>] [-c <size>] [-k seconds] [-d <level>] <port>\n"
+            + "\tcorfu_server (-l <path>|-m) [-fs] [-a <address>] [-t <token>] [-c <size>] [-k seconds] [-d <level>] <port>\n"
             + "\n"
             + "Options:\n"
             + " -l <path>, --log-path=<path>            Set the path to the storage file for the log unit.\n"
             + " -s, --single                            Deploy a single-node configuration.\n"
             + "                                         The server will be bootstrapped with a simple one-unit layout.\n"
+            + " -f, --sync                              Flush all writes to disk before acknowledging.\n"
             + " -a <address>, --address=<address>       IP address to advertise to external clients [default: localhost].\n"
             + " -m, --memory                            Run the unit in-memory (non-persistent).\n"
             + "                                         Data will be lost when the server exits!\n"
