+++ src/test/java/org/corfudb/AbstractCorfuTest.java
@@ -132,34 +132,4 @@ public class AbstractCorfuTest {
         }
 
     }
-
-    @Test
-    public void concurrentTestsExecute()
-    throws Exception
-    {
-        final AtomicLong l = new AtomicLong();
-        scheduleConcurrently(5, t -> l.getAndIncrement());
-        executeScheduled(5, 1000, TimeUnit.MILLISECONDS);
-        assertThat(l.get())
-                .isEqualTo(5);
-    }
-
-    @Test
-    public void concurrentTestsThrowExceptions()
-            throws Exception
-    {
-        scheduleConcurrently(5, t -> {throw new IOException("hi");});
-        assertThatThrownBy(() -> executeScheduled(5, 1000, TimeUnit.MILLISECONDS))
-                .isInstanceOf(IOException.class);
-    }
-
-    @Test
-    public void concurrentTestsTimeout()
-            throws Exception
-    {
-        scheduleConcurrently(5, t -> Thread.sleep(10000));
-        assertThatThrownBy(() -> executeScheduled(5, 1, TimeUnit.MILLISECONDS))
-                .isInstanceOf(CancellationException.class);
-    }
-
 }
